Remove user mode
User mode is currently broken, so we'd better remove it. It should be
added again later as part of issue #8.

The problem with the current implementation is that it does not retrieve
the source closure (.drv files) but the output closure from the nix
store. The latter is useless for security scanning.

Re #1
Re #8