Skip booted-system GC root during system check
After a bit of discussion, we decided that using the booted system as
additional GC root in system mode has more disadvantages than
advantages. The most important reason is that fixed packages don't
disappear from vulnix output listing until the next reboot. This is
clearly not practical especially in a server environment.

The most prominent downside is that an updated, but not booted kernel
will disappear from the listing.