Guess applied CVE patches from derivation file
Inspect the derivation and search for patches named `CVE-YYYY-NNNNN`.
vulnix assumes that these CVE have been fixed in the current version. If
a single patch fixes several CVEs at once, construct a file name that
names all of them.

Of course, patches can be removed if a new program version gets released
which is not known to be affected according to the current CVE database.
So we don't need to keep CVE patches forever. The whole mechanism is
only for fixes which don't change the version number. See `unzip` for
example.

Related to https://github.com/NixOS/nixpkgs/issues/15660