Basic whitelist match logic
Whitelist.filter() splits a list of derivations into two: first those
not covered by any whitelist rule, second a list of `Masked` objects
which state which of the derivations is covered by which whitelist item.
This gives us a bit of flexibility for user presentation.

Work in progress: tool is overall in an unusable state

Re #29313