Robust against mutiple version formats provided in the NVD (#77)
The NVD occasionally has inconsistently structured entries, which Vulnix generally tries to keep going through. #68 recently fixed one of the bugs which can arise from this, however, sometimes entries have multiple version formats which currently causes crashes.

One version format results in e.version being set to a string, whereas the other relies on it being a list. Calling append on a string causes the crash, which can be avoided by simply ensuring only one or the other format for version data is extracted. It isn't too important which is extracted, since this are generally temporary inconsistencies which are fixed on the NVD, so this needs to just be good enough and not crash.