Expose descriptions provided by the NVD
The NVD provides descriptions for CVEs, which are collected by Vulnix when retrieving vulnerability data. However,
Vulnix ignores this information, and doesn't expose it to the user. Whilst users can go online to read descriptions of
each vulnerability individually, this can become difficult if many Nix packages are installed, and so there are many
different CVEs. This information can be crucial for determining the real impact of a CVE given how a tool may be used,
and thus quickly isolating concerning CVEs for a system.

The descriptions can always be given in any JSON output as scripts handling this can easily ignore it if is not
relevant, but is behind a flag for human-readable output as descriptions can make the output much longer, so may
hinder current workflows.