Implement support for globs in a derivation's name/version
First of all, what I'm trying to solve: in most cases where I whitelist
a derivation, I'd like to not have it pinned down to a fixed version
(e.g. `php-calendar-7.4.30`) since I'd have to update the whitelist on
(almost) each nixpkgs update. OTOH only using the `pname` as sole
indicator is IMHO not enough in several cases, to give a few examples:

* `go-*-linux-amd64-bootstrap`: to exclude *any* Go CVE for the
  bootstrap compiler.
* `gitlab-14.*`: to exclude e.g. CVE-2021-22196 from gitlab 14 since the
  issue is only exploitable on 13.x.

The approach is relatively simple by using python's `fnmatch`[1] to
check if a glob-string matches a derivation's name in
`WhitelistRule#covers`, i.e.:

    >>> from fnmatch import fnmatch
    >>> fnmatch('gitlab-13.5', 'gitlab-14.*')
    False
    >>> fnmatch('gitlab-14.1', 'gitlab-14.*')
    True

Unfortunately this was non-trivial and contains a breaking change.
First, let me give some context:

* We currently have the map the string `${pname}-${version}` to `name`,
  a mandatory property of each derivation in Nix.
* When splitting a derivation's name (`builtins.parseDrvName`), Nix
  splits this into `name`/`version` on the first dash that's followed by
  a number.

This raises the question where to split e.g.
`go-*-linux-amd64-bootstrap` (to match `go-1.16-linux-amd64-bootstrap`).
The `pname` here is `go` and the `version` is `*-linux-amd64-bootstrap`,
however `*` is not a digit and thus the logic from
`builtins.parseDrvName` doesn't know where to split and considers the
entire string to be the derivation's name and the version to be empty.

On a personal note, these kinds of ambiguities we face here are the
reason why I'm not too happy about this `pname`/`version` thing.

To describe the entire change in detail:

* `WhitelistRule` now takes a single `name`-argument rather than
  `pname`/`version`.
* The `pname`/`version` of a `Deriv`-instance will be converted to
  `{pname}-{version}` in `WhitelistRule#covers` and then checked with
  `fnmatch` as shown above.
* Since the check whether a rule applies to a derivation is now done
  entirely in `WhitelistRule#covers`, I decided to remove
  `Whitelist#candidates` and just check each rule for each derivation.

This implies the following breaking changes:

* The YAML whitelists don't need a `version`-argument anymore, but
  instead rules are supposed to be specified in the same format as it's
  done in TOML, that means

      - name: libxslt
        version: '2.0'

  must become

      - name: libxslt-2.0

* The TOML (and also YAML) selector `{pname}` doesn't match a derivation
  in the schema `{pname}-{version}` anymore. That means

      ["libfoo"]
      cve = ...

  must become

      ["libfoo-*"]
      cve = ...

These can certainly be solved by checking both `libfoo` and `libfoo-*`
in `WhitelistRule#covers` and by concatenating `name` and `version` in
the YAML loader, but this is something I left out on purpose because I'd
like to get some feedback on the overall approach first.

[1] https://docs.python.org/3/library/fnmatch.html