Add CLI flag `-u` to show unused whitelist entries
When a deployment starts to age, it's easy to have things bitrotting in
the whitelists. This should be less of a big deal with the previous
change where globbing is implemented to make sure that whitelists affect
all derivations matching a certain glob pattern, but it's still a
problem from time to time.

The change is rather simple: each rule gets a counter that's increased
whenever `WhitelistRule#covers` returns `True`, i.e. when the rule takes
effect and silences an advisory. If `-u` is passed, each rule with a
counter equal to `0` will be dumped as JSON or in a TOML-like syntax to
stdout below the other output.

Fixes #31