Remove section header validation
This doesn't work anymore because e.g. `w3m-0.5.3+git20190105` is
perfectly valid, but not matched by the formatting regex. The testcase
states that this is to make sure that unintended whitespaces don't
confuse the parser, however I cannot reproduce that:

    $ python
    Python 3.9.13 (main, May 17 2022, 14:19:07)
    [GCC 11.3.0] on linux
    Type "help", "copyright", "credits" or "license" for more information.
    >>> import toml
    >>> toml.loads('[ "foo" ]')
    {'foo': {}}

I guess this was to circumvent an older bug and is just not relevant
anymore. Regarding the other testcase: I think that with the recently
added `-u` option it should be possible to find out a bit quicker why
the (probably unintended) whitespace between `pname`/`version` makes the
rule ineffective.

We may want to add a nicer implementation here (e.g. `--explain` that
states why a whitelist rule was matched), but for now I think that the
branch already contains enough changes ;-)